Remove DiskReader verify_close attribute
Any attempt to iterate over an instance of DiskReader will ensure that the
obj_size has been validated against the Content-Length in the metadata.  The
object-server already did this, and it is now enforced in tests as well.

When closing a DiskReader, if the _obj_size has been set (either by an
explicit call to get_obj_size, or by iterating over the body) the close method
will always automatically attempt to verify and possibly quarantine the file.
Since a Content-Length mis-match would have been caught before iterating the
file this always automatic close verification will only effect the ETag
validation which is only performed when the entire body of the DiskReader is
consumed.  Setting a "verify_close" attribute on the reader instance will have
no effect.

Remove the verify_file keyword argument to the DiskReader "close" method.

Change-Id: Id6a635842651de7807b9f1af348a88290107fd30