Container Confict on obj update with wrong Storage-Policy-Index
The container server knows its storage policy index by looking in the
container_stat table.  When an object update comes in to the container server,
there is an X-Storage-Policy-Index header in the incoming request, or it's
implicitly 0 during an upgrade. If the submitted X-Storage-Policy-Index
doesn't match the container's policy index, reject the request with a 409
Conflict response.

By doing this, we ensure that we know the policy of every object in a
container: it's the same as the container. If it doesn't match, it doesn't get
in. This makes reconciling two disagreeing container DBs possible.

Additionally we encode the policy index into the pending file to maintain
speed in put_object without having to acctually open the database and check
the storage policy in the stats table.  Benchmarking showed the extra get_info
call in put_object slowed the operation down by ~70%.  In this implementation
the storage_policy_index is only checked when a) creating the .pending_file
that doesn't already exist and b) merging the .pending_file.  Since the file
is truncated at merge (rather than removed) this is a onetime setup for the
database before put_object can start streaming in.  This change requires
_commit_puts to know the location of the pending file, and a base
implementation to move the .pending_file attribute to a property was added to
the common DatabaseBroker.  In the base implementation it's just a passthrough
to the existing base ._pending_file attribute.  For the ContainerBroker every
time commit_puts is called we poll the stats table to lookup the
storage_policy_index.  Because the commit_puts call is expected to hit the
stats table anyway during triggers this change was expected to not have
noticable impact on merge throughput and indeed benchmarking confirmed.

Implements: blueprint storage-policies
Change-Id: I3576f830f502229065d48f03ae9f45e2331aa8fa