Handle storage policies in container replicator
The container replicator will now sync up storage policy between two
different container DBs. The policy from the older container is kept,
or if the containers are the same age, a tiebreaker of lesser-policy
is used (this seems unlikely to occur, but we need some kind of
tiebreaker just in case).

The reason for keeping the older container's policy is that policy
mismatches are most likely to occur when a container is unavailable
for a short span and the application recreates it. When this happens,
the older container DB is likely to contain far more objects than the
newer one, so by picking the older, we minimize the number of
misplaced objects.

Change-Id: I0878fb32f30effd24eb791f6bed62a6045a2e8e5