Expose container info on deleted containers.
Include basic container info headers (X-Container-Object-Count, X-Timestamp,
etc.) on 404 responses from the container server.  Default empty values are
used as placeholders if the database does not exist.

Specifically the X-Backend-Status-Changed-At, X-Backend-DELETE-Timestamp and
the future X-Storage-Policy-Index value will be needed by the reconciler to
deal with reconciling out of order object writes in the face of recently
deleted containers.

 * Add "status_changed_at" key to the response from ContainerBroker.get_info.
 * Add "Status Timestamp" field to swift.cli.info.print_db_info_metadata.
 * Add "status_changed_at" key to the response from AccountBroker.get_info.

Conflicts:
	swift/container/backend.py
	swift/container/server.py
	test/unit/account/test_backend.py

Change-Id: Ie6d388f067f5b096b0f96faef151120ba23c8748