Add object-reconciler daemon
This daemon will take objects that are in the wrong storage policy and move
them to the right ones, or delete requests that went to the wrong storage
policy and apply them to the right ones. It operates on a queue similar to the
object-expirer's queue.

Discovering that the object is in the wrong policy will be done in subsequent
commits by the container replicator and the object updater; this is the daemon
that handles them once they happen.

Like the object expirer, you only need to run one of these per cluster see
etc/container-reconciler.conf.

Change-Id: I5ea62eb77ddcbc7cfebf903429f2ee4c098771c9