Convert POLICIES to list for random.choice
python's random.choice is implemented by taking a random index into the length
of a list and selecting the item at that integer index.

POLICIES is not a list, but happened mostly fit the interface for
random.choice since it had a len and supported integer indexing, but would
fail if your storage policy index's had any gaps.

In python3 a dict's .values() method returns a view instead of a real
list so changing StoragePolicyCollections __iter__ method to return .values()
instead of .itervalues() wouldn't buy us much.  It seems better to just
convert POLICIES to a list and conform to the contract for random.choice.

Change-Id: Id72a025822ecb6c8fec6b7be377d58ce9cb8f7b5