Keep status_changed_at up-to-date with status changes.
In particular during container recreation and replication.

When a container-server receives a PUT for a deleted database an extra
UPDATE is issued against the container_stat table to notate the
x-timestamp of the request.

During replication if merge_timestamps causes a container's status to
change (from DELETED to ACTIVE or vice-versa) the status_changed_at
filed is set to the current time.

Accurate reporting of status_changed_at is useful for container
replication forensics and allows resolution of "set on create"
attributes like the upcoming storage_policy_index.

Change-Id: I938919eb0dde7f473a18c32d1cb26986c1c3c5b7