Erasure Code Reconstructor
Here's a run down of the big things in here:

- addition of reconstruct_fa() function called from ssync_sender that
performs the actual reconstruction while sending the object over to
the receiver

- guts of reconstructor sorta follow replicator but there's not a
notiong of update() and upate_deleted().  There is a single job
processor and jobs are processed partition by partition.  At the
end of processing a partition, the ECrecon will try to remove it
and it may or may not work depending on whether everything was
deleted or not (conditionally on whether the job was moving data
or syncing)

- various ssync changes that correspond to what the ECRecon is doing
and how multi-FI support is built in but most of those are on a
different patch

There are still opportunities to add unit tests

Co-Authored-By: Alistair Coles <alistair.coles@hp.com>
Co-Authored-By: Thiago da Silva <thiago@redhat.com>
Co-Authored-By: Clay Gerrard <clay.gerrard@gmail.com>
Co-Authored-By: Tushar Gohad <tushar.gohad@intel.com>
Co-Authored-By: Samuel Merritt <sam@swiftstack.com>
Co-Authored-By: Christian Schwede <christian.schwede@enovance.com>
Co-Authored-By: Yuan Zhou <yuan.zhou@intel.com>

blueprint ec-reconstructor
Change-Id: I7d15620dc66ee646b223bb9fff700796cd6bef51